name not updating and whole table being deleted

+++ db/bounties.sql
@@ -4,6 +4,6 @@ CREATE TABLE bounties (
   id SERIAL8,
   name VARCHAR(255),
   species VARCHAR(255),
-  value INT8,
+  value INT,
   danger_level VARCHAR(255)
 );
